requirements.txt
Pillow
genanki
